# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: TaskFlow API Documentation
  version: 1.0.0
  description: >
    REST API for the TaskFlow student project. All successful API responses use
    `{ statusCode, success, message, data }`; errors use
    `{ statusCode, success, message, errors }`.
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server
...
paths:
...
  /projects:
    get:
      tags:
        - Projects
      summary: List user projects
      description: >
        **Purpose**: Retrieves all projects where the authenticated user is a registered member, including total member count and current user role.<br>
        **Authorization**: Required (`Bearer JWT`).<br>
        **Side Effects**: None.<br>
        **Expected Behavior**: Returns HTTP 200 with array of project membership objects.
      operationId: listProjects
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of projects retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: Projects fetched successfully!!!
                data:
                  - role: admin
                    project:
                      _id: 650c1f1e2f8b9a001c2d3e4f
                      name: FinTech Platform Modernization
                      description: Upgrading microservices architecture and implementing OpenAPI documentation.
                      members: 4
                      createdBy: 650c1f1e2f8b9a001c2d3e4f
                      createdAt: "2026-01-10T10:00:00.000Z"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Projects
      summary: Create a new project
      description: >
        **Purpose**: Initializes a new project workspace.<br>
        **Authorization**: Required (`Bearer JWT`).<br>
        **Side Effects**: Creates Project document in MongoDB and creates ProjectMember document assigning creator as `admin`.<br>
        **Expected Behavior**: Returns HTTP 201 with created Project object.
      operationId: createProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        "201":
          description: Project created and administrator role assigned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 201
                success: true
                message: Project created successfully
                data:
                  _id: 650c1f1e2f8b9a001c2d3e4f
                  name: Mobile App Development
                  description: Developing native iOS and Android applications for task management workflow.
                  createdBy: 650c1f1e2f8b9a001c2d3e4f
                  createdAt: "2026-01-16T11:20:00.000Z"
                  updatedAt: "2026-01-16T11:20:00.000Z"
        "400":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /projects/{projectId}:
    get:
      tags:
        - Projects
      summary: Get project by ID
      description: >
        **Purpose**: Fetches detailed metadata for a specific project.<br>
        **Authorization**: Required (`Bearer JWT`). User must be a member (`admin`, `project_admin`, `member`).<br>
        **Side Effects**: None.<br>
        **Expected Behavior**: Returns HTTP 200 with Project object. Throws HTTP 403 if user is not a project member.
      operationId: getProjectById
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the target project
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
      responses:
        "200":
          description: Project details fetched successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: Project fetched successfully!!!
                data:
                  _id: 650c1f1e2f8b9a001c2d3e4f
                  name: FinTech Platform Modernization
                  description: Upgrading microservices architecture and implementing OpenAPI documentation.
                  createdBy: 650c1f1e2f8b9a001c2d3e4f
                  createdAt: "2026-01-10T10:00:00.000Z"
                  updatedAt: "2026-01-10T10:00:00.000Z"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Projects
      summary: Update project details
      description: >
        **Purpose**: Updates name or description of an existing project.<br>
        **Authorization**: Required (`Bearer JWT`). Role restricted to `admin` only.<br>
        **Side Effects**: Updates Project document in MongoDB.<br>
        **Expected Behavior**: Returns HTTP 200 with updated Project details.
      operationId: updateProject
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the project
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        "200":
          description: Project updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: Project details updated successfully!!!
                data:
                  _id: 650c1f1e2f8b9a001c2d3e4f
                  name: FinTech Platform & Mobile Modernization
                  description: Upgrading microservices architecture, implementing OpenAPI specs, and launching iOS app.
                  createdBy: 650c1f1e2f8b9a001c2d3e4f
                  createdAt: "2026-01-10T10:00:00.000Z"
                  updatedAt: "2026-01-16T14:00:00.000Z"
        "400":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Projects
      summary: Delete a project
      description: >
        **Purpose**: Permanently deletes a project workspace and all child entities.<br>
        **Authorization**: Required (`Bearer JWT`). Role restricted to `admin` only.<br>
        **Side Effects**: Cascade deletes Project, ProjectMember, Task, SubTask, and Note records from MongoDB.<br>
        **Expected Behavior**: Returns HTTP 200 acknowledging project deletion.
      operationId: deleteProject
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the project
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
      responses:
        "200":
          description: Project and associated resources cascade deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: Project deleted successfully!!!
                data: null
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /projects/{projectId}/members:
    get:
      tags:
        - Projects
      summary: Get project members
      description: >
        **Purpose**: Retrieves list of members and assigned roles in a project workspace.<br>
        **Authorization**: Required (`Bearer JWT`). User must be a member (`admin`, `project_admin`, `member`).<br>
        **Side Effects**: None.<br>
        **Expected Behavior**: Returns HTTP 200 with populated member list containing user details (username, email, avatar).
      operationId: getProjectMembers
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the project
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
      responses:
        "200":
          description: Project members list retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: Project members fetched successfully
                data:
                  - _id: 650c1f1e2f8b9a001c2d3e50
                    user:
                      _id: 650c1f1e2f8b9a001c2d3e4f
                      username: alex_rivera
                      email: alex.rivera@taskflow.dev
                      avatar:
                        url: https://placehold.co/200?text=No+Image
                    project: 650c1f1e2f8b9a001c2d3e4f
                    role: admin
                  - _id: 650c1f1e2f8b9a001c2d3e51
                    user:
                      _id: 650c1f1e2f8b9a001c2d3e99
                      username: sam_dev
                      email: sam.dev@taskflow.dev
                      avatar:
                        url: https://placehold.co/200?text=No+Image
                    project: 650c1f1e2f8b9a001c2d3e4f
                    role: member
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Projects
      summary: Add member to project
      description: >
        **Purpose**: Invites and adds a registered user to the project workspace.<br>
        **Authorization**: Required (`Bearer JWT`). Role restricted to `admin` and `project_admin`.<br>
        **Side Effects**: Creates ProjectMember record and dispatches email notification with project access link.<br>
        **Expected Behavior**: Returns HTTP 201 with created ProjectMember object. Throws HTTP 409 if user is already a member.
      operationId: addProjectMember
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the target project
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProjectMemberRequest'
      responses:
        "201":
          description: User added to project and notification email sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 201
                success: true
                message: Member added successfully
                data:
                  _id: 650c1f1e2f8b9a001c2d3e51
                  user: 650c1f1e2f8b9a001c2d3e99
                  project: 650c1f1e2f8b9a001c2d3e4f
                  role: member
        "400":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /projects/{projectId}/members/{userId}:
    put:
      tags:
        - Projects
      summary: Update project member role
      description: >
        **Purpose**: Updates the role permissions (`admin`, `project_admin`, `member`) of an existing team member.<br>
        **Authorization**: Required (`Bearer JWT`). Restricted to project `admin` only.<br>
        **Side Effects**: Updates `role` on ProjectMember document in MongoDB.<br>
        **Expected Behavior**: Returns HTTP 200 with updated ProjectMember object.
      operationId: updateMemberRole
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the project
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
        - name: userId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the user member
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e99
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRoleRequest'
      responses:
        "200":
          description: Member role updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: Role updated succesfully!!!
                data:
                  _id: 650c1f1e2f8b9a001c2d3e51
                  user: 650c1f1e2f8b9a001c2d3e99
                  project: 650c1f1e2f8b9a001c2d3e4f
                  role: project_admin
        "400":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Projects
      summary: Remove member from project
      description: >
        **Purpose**: Removes a user from a project workspace.<br>
        **Authorization**: Required (`Bearer JWT`). Restricted to `admin` and `project_admin` roles.<br>
        **Side Effects**: Deletes ProjectMember record from MongoDB.<br>
        **Expected Behavior**: Returns HTTP 200 with removed ProjectMember details.
      operationId: removeProjectMember
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the project
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e4f
        - name: userId
          in: path
          required: true
          description: 24-character hex MongoDB ObjectId of the user member to remove
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 650c1f1e2f8b9a001c2d3e99
      responses:
        "200":
          description: Member removed from project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                success: true
                message: Member removed successfully!!!
                data:
                  _id: 650c1f1e2f8b9a001c2d3e51
                  user: 650c1f1e2f8b9a001c2d3e99
                  project: 650c1f1e2f8b9a001c2d3e4f
                  role: member
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter valid JWT Access Token obtained from `/api/v1/auth/login`.
  schemas:
...
    CreateProjectRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          minLength: 5
          example: FinTech Platform Modernization
          description: Project name (minimum 5 characters)
        description:
          type: string
          minLength: 20
          example: Upgrading microservices architecture and implementing OpenAPI documentation.
          description: Detailed project description (minimum 20 characters)
    UpdateProjectRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          minLength: 5
          example: FinTech Platform & Mobile Modernization
          description: Updated project name
        description:
          type: string
          minLength: 20
          example: Upgrading microservices architecture, implementing OpenAPI specs, and launching iOS app.
          description: Updated project description
    AddProjectMemberRequest:
      type: object
      required:
        - email
        - role
      properties:
        email:
          type: string
          format: email
          example: sam.dev@taskflow.dev
          description: Registered user email to invite
        role:
          type: string
          enum: [admin, project_admin, member]
          example: member
          description: Project permission role assigned to invited member
    UpdateMemberRoleRequest:
      type: object
      required:
        - newRole
      properties:
        newRole:
          type: string
          enum: [admin, project_admin, member]
          example: project_admin
          description: Updated role permission for team member
...
    ErrorResponse:
      type: object
      required:
        - statusCode
        - success
        - message
        - errors
      properties:
        statusCode:
          type: integer
          example: 400
          description: HTTP response status code
        success:
          type: boolean
          example: false
          description: Success flag (false for errors)
        message:
          type: string
          example: Request failed due to invalid input validation.
          description: Error message summary
        errors:
          type: array
          description: Array of detailed validation error objects or strings
          items:
            type: object
          example:
            - field: email
              message: Invalid email format
        stack:
          type: string
          nullable: true
          example: "Error: Bad Request at..."
          description: Stack trace snippet (only populated in development environment)
    SuccessResponse:
      type: object
      required:
        - statusCode
        - success
        - message
        - data
      properties:
        statusCode:
          type: integer
          example: 200
          description: HTTP response status code
        success:
          type: boolean
          example: true
          description: Success flag (true for successful responses)
        message:
          type: string
          example: Operation completed successfully
          description: Success message summary
        data:
          nullable: true
          description: Response data payload. Its exact type is documented for every endpoint example and in FRONTEND_API_CONTRACT.md.
...
  responses:
    Unauthorized:
      description: Unauthorized - JWT Access Token is missing, invalid, or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 401
            success: false
            message: Unauthorized request or invalid JWT token
            errors: []
    Forbidden:
      description: Forbidden - Insufficient role permissions to access this endpoint or project workspace.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 403
            success: false
            message: You do not have permission to perform this action
            errors: []
    NotFound:
      description: Not Found - The requested resource (user, project, task, subtask, note) does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 404
            success: false
            message: Resource not found
            errors: []
    Conflict:
      description: Conflict - Operation conflicts with existing state (e.g. duplicate username, email, or existing project member).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 409
            success: false
            message: User with this email or username already exists
            errors: []
...
    InternalServerError:
      description: Internal Server Error - An unexpected server error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 500
            success: false
            message: Internal Server Error
            errors: []
    ValidationError:
      description: Validation Error - Input parameters or payload failed express-validator rules.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 400
            success: false
            message: Validation Error
            errors:
              - field: email
                message: Invalid email format